src/auth-service/requirements.txt
Flask==2.3.2
PyJWT==2.7.0
psycopg2-binary==2.9.6
flask-cors==4.0.0
